pysrc/requirements.txt
altgraph==0.17
aniso8601==8.0.0
Boruta==0.3
click==7.1.2
cycler==0.10.0
dnspython==2.0.0
docx==0.2.4
eventlet==0.25.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.0
gevent==20.5.1
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
imbalanced-learn==0.6.2
imblearn==0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.2
macholib==1.14
MarkupSafe==1.1.1
matplotlib==3.2.2
missingpy==0.2.0
monotonic==1.5
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2020.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
tqdm==4.45.0
Werkzeug==1.0.1
xlrd==1.2.0
zope.event==4.4
zope.interface==5.1.0
